+++ __tests__/ContentArea.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { ContentArea } from '../ContentArea';
+import { ContentArea } from '../frontend/src/components/ContentArea';
 
 describe('ContentArea Component', () => {
   const mockOnFilterChange = vi.fn();
