+++ __tests__/FilterBar.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { FilterBar } from '../FilterBar';
+import { FilterBar } from '../frontend/src/components/FilterBar';
 
 describe('FilterBar Component', () => {
   const mockOnFilterChange = vi.fn();
