+++ __tests__/passwordValidator.test.ts
